+++ api/__init__.py
@@ -1,0 +1,2 @@
+"""Implements API Queries to a LAPIS servers as CovSpectrum, 
+   out our wastewater Wise-Loculus."""
